fix: 로그아웃 에러 해결
- 카카오 토큰이 이미 만료되었을 경우 로그아웃 성공 처리
- kakao-proxy-server의 session 유틸리티를 shared/auth 패키지로 이동
- 공통 에러 상수를 shared/auth/constants에 추가
- 인증 관련 타입 정의를 shared 패키지로 통합
- 코드 재사용성 및 유지보수성 향상
