+++ apps/kakao-proxy-server/package.json
@@ -25,6 +25,7 @@
         "morgan": "^1.10.0"
     },
     "devDependencies": {
+        "@types/cookie-parser": "^1.4.9",
         "@types/cors": "^2.8.17",
         "@types/express": "^5.0.0",
         "@types/morgan": "^1.9.9",
