+++ packages/shared/auth/src/constants/error.ts
@@ -1,0 +1,3 @@
+export const ERROR_MSG = {
+    SESSION_SECRET_NOT_SET: 'SESSION_SECRET 환경 변수가 설정되지 않았습니다.',
+};
