+++ packages/shared/auth/src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './token';
+export * from './error';
